Added two ultimate abilities 1: Multi bullet shot 2: Invincibility
Added two more sprites for both abilities

I tested this extensively, but knowing grover he will break it in three seconds.
Fixed bug that sometimes made multi bullet disappear after pressing Q moments after shooting a regular bullet.

Instead of making another class for the multi bullet, I modified the Bullet class to accept a boolean whether or not it is a regular bullet or not.