Added a couple of extra features.
Added the randomized movement of enemies which is triggered when the enemy_count is half the original enemy count. Also added the extra ship which flies from left to right then right to left once per level. Lastly, I added some more enemies which are chosen at random. Still working on getting them to properly appear which background. One thing I added which should be noted is the new "state" for enemy's which dictates whether they move at random or in sync.